requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy==1.19.2
pandas==1.1.3
matplotlib==3.3.2
seaborn==0.11.0
scikit-image==0.17.2
tqdm==4.50.2
pyyaml==5.3.1
pillow==8.0.1